change index to a static site generated by gulp-mustache
Converted over the json file to a categorised data format, added a template index file to generate from. Added support for ads.
