+++ ads.txt
@@ -1,0 +1,1 @@
+google.com, pub-8168184751026221, DIRECT, f08c47fec0942fa0
